+++ modules/apps/account/account-api/src/main/java/com/liferay/account/service/AccountEntryLocalServiceUtil.java
@@ -331,8 +331,8 @@ public class AccountEntryLocalServiceUtil {
 	public static com.liferay.portal.kernel.search.BaseModelSearchResult
 		<com.liferay.account.model.AccountEntry> search(
 			long companyId, String keywords,
-			java.util.LinkedHashMap<String, Object> params, long cur,
-			long delta, String orderByField, boolean reverse) {
+			java.util.LinkedHashMap<String, Object> params, int cur, int delta,
+			String orderByField, boolean reverse) {
 
 		return getService().search(
 			companyId, keywords, params, cur, delta, orderByField, reverse);
